+++ src/js/theme.js
@@ -3,8 +3,3 @@ export * from "./search.js";
 export * from "./mobile.js";
 export * from "./tinyDrawer.js";
 export * from "./leader-line.js";
-
-// Accept proper HMR in Parcel
-if (module && module.hot) {
-  module.hot.accept();
-}